Dev: create menu before QuestionActivity.kt
