+++ app/src/main/res/layout/activity_main.xml
@@ -7,82 +7,36 @@
     tools:context=".MainActivity">
 
     <TextView
-        android:id="@+id/trueOption"
-        style="@style/TextAppearance.AppCompat.Display1"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:background="@color/teal_200"
-        android:clickable="true"
-        android:elevation="4dp"
-        android:focusable="true"
-        android:foreground="?attr/selectableItemBackground"
-        android:gravity="center"
-        android:padding="16dp"
-        android:text="TRUE"
-        android:textStyle="bold"
-        app:layout_constraintBottom_toTopOf="@+id/question"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:text="False – at a low orbit the Great Wall is visible, but no man-made structure is visible from outer space or the moon." />
-
-    <TextView
-        android:id="@+id/falseOption"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:padding="16dp"
-        android:background="@android:color/holo_red_light"
-        android:gravity="center"
-        android:text="FALSE"
-        android:foreground="?attr/selectableItemBackground"
-        android:clickable="true"
-        android:focusable="true"
-        style="@style/TextAppearance.AppCompat.Display1"
-        android:textStyle="bold"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/question"
-        android:elevation="4dp"/>
-
-    <TextView
-        android:id="@+id/question"
+        android:id="@+id/title"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:paddingStart="16dp"
-        android:paddingTop="20dp"
-        android:paddingEnd="16dp"
-        android:paddingBottom="20dp"
-        android:text="Hello World!"
-        android:textStyle="bold"
-        android:gravity="center"
-        style="@style/TextAppearance.AppCompat.Headline"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-    <TextView
-        android:id="@+id/counter"
-        android:text="1/5"
+        android:text="[CrossFire ❌🔥]"
+        android:textSize="24dp"
+        android:textColor="@color/black"
+        app:layout_constraintBottom_toTopOf="@id/playButton"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintVertical_chainStyle="packed" />
+
+    <Button
+        android:id="@+id/playButton"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        app:layout_constraintEnd_toEndOf="@id/loader"
-        app:layout_constraintStart_toStartOf="@id/loader"
-        app:layout_constraintTop_toTopOf="@id/loader"
-        app:layout_constraintBottom_toBottomOf="@id/loader"
-        android:elevation="4dp"
-        />
-    <ProgressBar
-        android:id="@+id/loader"
-        android:indeterminateTint="@color/cardview_dark_background"
-        android:layout_width="64dp"
-        android:layout_height="64dp"
-        android:layout_margin="8dp"
-        android:indeterminate="true"
-        android:visibility="invisible"
-        android:indeterminateBehavior="repeat"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:elevation="4dp"/>
+        android:text="Play"
+        app:layout_constraintBottom_toTopOf="@id/exitButton"
+        app:layout_constraintTop_toBottomOf="@id/title"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"/>
+
+    <Button
+        android:id="@+id/exitButton"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Exit"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/playButton"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"/>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
